requirements.txt
# AI Real Estate Marketing System - Dependencies

# Core
streamlit>=1.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0

# AI & ML
openai>=1.3.0
elevenlabs>=1.0.0
replicate>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0
controlnet-aux>=0.0.7
opencv-python>=4.8.0

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# API & Utilities
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
python-multipart>=0.0.6

# Database & Auth
supabase>=2.0.0
clerk-sdk>=2.0.0

# UI Enhancements
streamlit-option-menu>=0.3.6
streamlit-webrtc>=0.47.0

# Logging & Monitoring
loguru>=0.7.0
sentry-sdk>=1.39.0

# Development
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0


